third_party/blink/web_tests/platform/generic/svg/dynamic-updates/SVGLinearGradientElement-svgdom-x2-prop-expected.txt

SVG 1.1 dynamic update tests

Tests dynamic updates of the 'x2' property of the SVGLinearGradientElement object

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS linearGradientElement.x2.baseVal.valueAsString is "-100%"
PASS linearGradientElement.x2.baseVal.valueAsString is "100%"
PASS successfullyParsed is true

TEST COMPLETE
